Major refactoring
Refactor much of -cfout- body (not subroutines):
1. Be conceptually clear and consistent about each varlist local's
definition, using this to implement clearer checks of differing variable
types.
2. Issue warnings only after checking that there are no errors.
3. Return more stored results: r(varlist), r(varonlym), r(difftype),
r(alldiff), r(Nonlym), and r(Nonlyu). The philosophy is that if a
variable or observation is mentioned in a warning, it should be
accessible via a stored result.
4. Do not reload the master dataset after modifying the using: merge
modified master into using.
5. Prepare for option -id()- to take a varlist.
6. Prepare to remove the use of the _cf prefix, using temporary variable
names instead.