Create the differences dataset much more quickly.
The new implementation does not resize the Mata matrices containing the
dataset. For a big dataset on my computer, the new version is running
about 10x faster. It skips the creation entirely if option -saving()- is
not specified. It also does not save Mata globals.

As a side benefit, the ID variable in the differences dataset is now not
uniformly str244.

Further, this version removes the use of the _cf prefix.